requirements.txt
Pillow
pymilvus
scikit_learn
streamlit==1.41.1 # Last Version that handles CUDA well
timm
torch
